feat: add codegen package
