+++ codegen/package/pkl.python@0.1.11
@@ -1,0 +1,17 @@
+{
+  "name": "pkl.python",
+  "packageUri": "package://github.com/jw-y/pkl-python/pkl.python@0.1.11",
+  "version": "0.1.11",
+  "packageZipUrl": "https://github.com/jw-y/pkl-python/releases/download/pkl.python@0.1.11/pkl.python@0.1.11.zip",
+  "packageZipChecksums": {
+    "sha256": "a52edebaf6b4366e098f9fdc62074bbabf8e2602d5f43875a0a4b3e877fdb7a5"
+  },
+  "dependencies": {},
+  "sourceCodeUrlScheme": "https://github.com/jw-y/pkl-python/blob/0.1.11/codegen/src%{path}#L%{line}-L%{endLine}",
+  "sourceCode": "https://github.com/jw-y/pkl-python",
+  "license": "Apache-2.0",
+  "authors": [
+    "Jungwoo Yang <jwyang0213@gmail.com>"
+  ],
+  "description": "Pkl bindings for the Python programming language"
+}
